+++ test-resources/babashka/tasks_cli_ns.clj
@@ -1,0 +1,10 @@
+(ns babashka.tasks-cli-ns
+  "Namespace-level :org.babashka/cli defaults, merged under fn metadata like bb -x."
+  {:org.babashka/cli {:restrict true
+                      :spec {:verbose {:coerce :boolean :desc "Verbose"}}}})
+
+(defn go
+  "Runs it"
+  {:org.babashka/cli {:spec {:port {:coerce :int :desc "Port"}}}}
+  [opts]
+  (prn (assoc opts :ran :go)))